--- opencti-platform/opencti-front/src/private/components/common/stix_core_relationships/SimpleStixObjectOrStixRelationshipStixCoreRelationshipLine.jsx
@@ -1,806 +1,0 @@
-import React, { Component } from 'react';
-import * as PropTypes from 'prop-types';
-import { compose } from 'ramda';
-import { Link } from 'react-router-dom';
-import { graphql, createFragmentContainer } from 'react-relay';
-import withStyles from '@mui/styles/withStyles';
-import withTheme from '@mui/styles/withTheme';
-import ListItem from '@mui/material/ListItem';
-import ListItemIcon from '@mui/material/ListItemIcon';
-import ListItemText from '@mui/material/ListItemText';
-import { MoreVertOutlined } from '@mui/icons-material';
-import { AutoFix } from 'mdi-material-ui';
-import Skeleton from '@mui/material/Skeleton';
-import Tooltip from '@mui/material/Tooltip';
-import * as R from 'ramda';
-import { ListItemButton } from '@mui/material';
-import Box from '@mui/material/Box';
-import { DraftChip, getDraftModeColor } from '../draft/DraftChip';
-import inject18n from '../../../../components/i18n';
-import ItemConfidence from '../../../../components/ItemConfidence';
-import StixCoreRelationshipPopover from './StixCoreRelationshipPopover';
-import Security from '../../../../utils/Security';
-import { KNOWLEDGE_KNUPDATE } from '../../../../utils/hooks/useGranted';
-import ItemIcon from '../../../../components/ItemIcon';
-import ItemMarkings from '../../../../components/ItemMarkings';
-import ItemEntityType from '../../../../components/ItemEntityType';
-
-const styles = (theme) => ({
-  item: {
-    height: 50,
-  },
-  itemIcon: {
-    color: theme.palette.primary.main,
-  },
-  bodyItem: {
-    height: 20,
-    fontSize: 13,
-    float: 'left',
-    whiteSpace: 'nowrap',
-    overflow: 'hidden',
-    textOverflow: 'ellipsis',
-    paddingRight: 10,
-  },
-  itemIconDisabled: {
-    color: theme.palette.grey[700],
-  },
-  placeholder: {
-    display: 'inline-block',
-    height: '1em',
-    backgroundColor: theme.palette.grey[700],
-  },
-});
-
-class SimpleStixObjectOrStixRelationshipStixCoreRelationshipLineComponent extends Component {
-  render() {
-    const {
-      fsd,
-      t,
-      classes,
-      theme,
-      dataColumns,
-      node,
-      paginationOptions,
-      entityId,
-      entityLink,
-      connectionKey,
-    } = this.props;
-    const link = `${entityLink}/relations/${node.id}`;
-    const isReversed = node.fromId === entityId;
-    const row = isReversed ? node.to : node.from;
-    const element = row || {
-      id: isReversed ? node.toId : node.fromId,
-      entity_type: 'Restricted',
-      restricted: true,
-      representative: { main: t('Restricted') },
-    };
-    const draftColor = getDraftModeColor(theme);
-    return (
-      <ListItem
-        divider={true}
-        disablePadding
-        secondaryAction={node.is_inferred ? (
-          <Tooltip
-            title={
-              t('Inferred knowledge based on the rule ')
-              + R.head(node.x_opencti_inferences).rule.name
-            }
-          >
-            <AutoFix fontSize="small" style={{ marginLeft: -30 }} />
-          </Tooltip>
-        ) : (
-          <Security needs={[KNOWLEDGE_KNUPDATE]}>
-            <StixCoreRelationshipPopover
-              stixCoreRelationshipId={node.id}
-              paginationOptions={paginationOptions}
-              connectionKey={connectionKey}
-            />
-          </Security>
-        )}
-      >
-        <ListItemButton
-          classes={{ root: classes.item }}
-          component={Link}
-          to={link}
-        >
-          <ListItemIcon classes={{ root: classes.itemIcon }}>
-            <ItemIcon type={node.entity_type} isReversed={isReversed} color={node.draftVersion ? draftColor : null} />
-          </ListItemIcon>
-          <ListItemText
-            primary={
-              <>
-                <div
-                  className={classes.bodyItem}
-                  style={{ width: dataColumns.relationship_type.width }}
-                >
-                  <ItemEntityType
-                    entityType={node.relationship_type}
-                  />
-                </div>
-                <div
-                  className={classes.bodyItem}
-                  style={{ width: dataColumns.entity_type.width }}
-                >
-                  <ItemEntityType
-                    entityType={element.entity_type}
-                    size='large'
-                    showIcon
-                    isRestricted={!row}
-                  />
-                </div>
-                <div
-                  className={classes.bodyItem}
-                  style={{ width: dataColumns.name.width }}
-                >
-                  {element.representative.main}
-                  {element.draftVersion && (<DraftChip/>)}
-                </div>
-                <div
-                  className={classes.bodyItem}
-                  style={{ width: dataColumns.created_at.width }}
-                >
-                  {fsd(node.created_at)}
-                </div>
-                <div
-                  className={classes.bodyItem}
-                  style={{ width: dataColumns.confidence.width }}
-                >
-                  <ItemConfidence
-                    confidence={node.confidence}
-                    entityType={node.entity_type}
-                    variant="inList"
-                  />
-                </div>
-                <div
-                  className={classes.bodyItem}
-                  style={{ width: dataColumns.markings.width }}
-                >
-                  <ItemMarkings
-                    variant="inList"
-                    markingDefinitions={node.objectMarking ?? []}
-                    limit={1}
-                  />
-                </div>
-              </>
-          }
-          />
-        </ListItemButton>
-      </ListItem>
-    );
-  }
-}
-
-SimpleStixObjectOrStixRelationshipStixCoreRelationshipLineComponent.propTypes = {
-  dataColumns: PropTypes.object,
-  entityId: PropTypes.string,
-  entityLink: PropTypes.string,
-  paginationOptions: PropTypes.object,
-  node: PropTypes.object,
-  classes: PropTypes.object,
-  t: PropTypes.func,
-  fsd: PropTypes.func,
-  connectionKey: PropTypes.string,
-};
-
-const SimpleStixObjectOrStixRelationshipStixCoreRelationshipLineFragment = createFragmentContainer(
-  SimpleStixObjectOrStixRelationshipStixCoreRelationshipLineComponent,
-  {
-    node: graphql`
-        fragment SimpleStixObjectOrStixRelationshipStixCoreRelationshipLine_node on StixCoreRelationship {
-          id
-          draftVersion {
-            draft_id
-            draft_operation
-          }
-          entity_type
-          parent_types
-          relationship_type
-          confidence
-          start_time
-          stop_time
-          description
-          is_inferred
-          created_at
-          x_opencti_inferences {
-            rule {
-              id
-              name
-            }
-          }
-          objectMarking {
-            id
-            definition_type
-            definition
-            x_opencti_order
-            x_opencti_color
-          }
-          from {
-            ... on StixDomainObject {
-              representative {
-                main
-              }
-              id
-              draftVersion {
-                draft_id
-                draft_operation
-              }
-              entity_type
-              parent_types
-              created_at
-              updated_at
-              objectLabel {
-                id
-                value
-                color
-              }
-            }
-            ... on AttackPattern {
-              name
-              description
-              x_mitre_id
-              killChainPhases {
-                id
-                phase_name
-                x_opencti_order
-              }
-              objectMarking {
-                id
-                definition_type
-                definition
-                x_opencti_order
-                x_opencti_color
-              }
-              objectLabel {
-                id
-                value
-                color
-              }
-            }
-            ... on Campaign {
-              name
-              description
-            }
-            ... on CourseOfAction {
-              name
-              description
-            }
-            ... on Individual {
-              name
-              description
-            }
-            ... on Organization {
-              name
-              description
-            }
-            ... on Sector {
-              name
-              description
-            }
-            ... on System {
-              name
-              description
-            }
-            ... on Indicator {
-              name
-              description
-            }
-            ... on Infrastructure {
-              name
-              description
-            }
-            ... on IntrusionSet {
-              name
-              description
-            }
-            ... on Position {
-              name
-              description
-            }
-            ... on City {
-              name
-              description
-            }
-            ... on AdministrativeArea {
-              name
-              description
-            }
-            ... on Country {
-              name
-              description
-            }
-            ... on Region {
-              name
-              description
-            }
-            ... on Malware {
-              name
-              description
-            }
-            ... on MalwareAnalysis {
-              result_name
-            }
-            ... on ThreatActor {
-              name
-              description
-            }
-            ... on Tool {
-              name
-              description
-            }
-            ... on Vulnerability {
-              name
-              description
-            }
-            ... on Incident {
-              name
-              description
-            }
-            ... on Event {
-              name
-              description
-            }
-            ... on Channel {
-              name
-              description
-            }
-            ... on Narrative {
-              name
-              description
-            }
-            ... on Language {
-              name
-            }
-            ... on DataComponent {
-              name
-            }
-            ... on DataSource {
-              name
-            }
-            ... on Case {
-              name
-            }
-            ... on Report {
-              name
-            }
-            ... on Grouping {
-              name
-            }
-            ... on Note {
-              attribute_abstract
-              content
-            }
-            ... on Opinion {
-              opinion
-            }
-            ... on ObservedData {
-              name
-            }
-            ... on StixCyberObservable {
-              id
-              entity_type
-              parent_types
-              observable_value
-              representative {
-                main
-              }
-              objectMarking {
-                id
-                definition_type
-                definition
-                x_opencti_order
-                x_opencti_color
-              }
-              objectLabel {
-                id
-                value
-                color
-              }
-            }
-            ... on Indicator {
-              id
-              name
-              pattern_type
-              pattern_version
-              description
-              valid_from
-              valid_until
-              x_opencti_score
-              x_opencti_main_observable_type
-              created
-              objectMarking {
-                id
-                definition_type
-                definition
-                x_opencti_order
-                x_opencti_color
-              }
-              objectLabel {
-                id
-                value
-                color
-              }
-            }
-            ... on BasicRelationship {
-              id
-              entity_type
-              parent_types
-            }
-            ... on StixCoreRelationship {
-              relationship_type
-              representative {
-                main
-              }
-              from {
-                ... on BasicObject {
-                  id
-                  entity_type
-                }
-                ... on BasicRelationship {
-                  id
-                }
-                ... on StixCyberObservable {
-                  representative {
-                    main
-                  }
-                }
-                ... on StixDomainObject {
-                  representative {
-                    main
-                  }
-                }
-              }
-              to {
-                ... on BasicObject {
-                  id
-                  entity_type
-                }
-                ... on BasicRelationship {
-                  id
-                }
-                ... on StixCyberObservable {
-                  representative {
-                    main
-                  }
-                }
-                ... on StixDomainObject {
-                  representative {
-                    main
-                  }
-                }
-              }
-            }
-          }
-          fromId
-          fromType
-          to {
-            ... on StixDomainObject {
-              id
-              representative {
-                main
-              }
-              draftVersion {
-                draft_id
-                draft_operation
-              }
-              entity_type
-              parent_types
-              created_at
-              updated_at
-              objectLabel {
-                id
-                value
-                color
-              }
-            }
-            ... on AttackPattern {
-              name
-              description
-              x_mitre_id
-              killChainPhases {
-                id
-                phase_name
-                x_opencti_order
-              }
-              objectMarking {
-                id
-                definition_type
-                definition
-                x_opencti_order
-                x_opencti_color
-              }
-              objectLabel {
-                id
-                value
-                color
-              }
-            }
-            ... on Campaign {
-              name
-              description
-            }
-            ... on CourseOfAction {
-              name
-              description
-            }
-            ... on Individual {
-              name
-              description
-            }
-            ... on Organization {
-              name
-              description
-            }
-            ... on Sector {
-              name
-              description
-            }
-            ... on System {
-              name
-              description
-            }
-            ... on Indicator {
-              name
-              description
-            }
-            ... on Infrastructure {
-              name
-              description
-            }
-            ... on IntrusionSet {
-              name
-              description
-            }
-            ... on Position {
-              name
-              description
-            }
-            ... on City {
-              name
-              description
-            }
-            ... on AdministrativeArea {
-              name
-              description
-            }
-            ... on Country {
-              name
-              description
-            }
-            ... on Region {
-              name
-              description
-            }
-            ... on Malware {
-              name
-              description
-            }
-            ... on MalwareAnalysis {
-              result_name
-            }
-            ... on ThreatActor {
-              name
-              description
-            }
-            ... on Tool {
-              name
-              description
-            }
-            ... on Vulnerability {
-              name
-              description
-            }
-            ... on Incident {
-              name
-              description
-            }
-            ... on Event {
-              name
-              description
-            }
-            ... on Channel {
-              name
-              description
-            }
-            ... on Narrative {
-              name
-              description
-            }
-            ... on Language {
-              name
-            }
-            ... on DataComponent {
-              name
-            }
-            ... on DataSource {
-              name
-            }
-            ... on Case {
-              name
-            }
-            ... on Report {
-              name
-            }
-            ... on Grouping {
-              name
-            }
-            ... on Note {
-              attribute_abstract
-              content
-            }
-            ... on Opinion {
-              opinion
-            }
-            ... on ObservedData {
-              name
-            }
-            ... on StixCyberObservable {
-              id
-              entity_type
-              parent_types
-              observable_value
-              representative {
-                main
-              }
-              objectMarking {
-                id
-                definition_type
-                definition
-                x_opencti_order
-                x_opencti_color
-              }
-              objectLabel {
-                id
-                value
-                color
-              }
-            }
-            ... on Indicator {
-              id
-              name
-              pattern_type
-              pattern_version
-              description
-              valid_from
-              valid_until
-              x_opencti_score
-              x_opencti_main_observable_type
-              created
-              objectMarking {
-                id
-                definition_type
-                definition
-                x_opencti_order
-                x_opencti_color
-              }
-              objectLabel {
-                id
-                value
-                color
-              }
-            }
-            ... on BasicRelationship {
-              id
-              entity_type
-              parent_types
-            }
-            ... on StixCoreRelationship {
-              relationship_type
-              representative {
-                main
-              }
-            }
-          }
-          toId
-          toType
-          killChainPhases {
-            id
-            phase_name
-            x_opencti_order
-          }
-        }
-      `,
-  },
-);
-
-export const SimpleStixObjectOrStixRelationshipStixCoreRelationshipLine = compose(
-  inject18n,
-  withStyles(styles),
-  withTheme,
-)(SimpleStixObjectOrStixRelationshipStixCoreRelationshipLineFragment);
-
-class SimpleStixObjectOrStixRelationshipStixCoreRelationshipLineDummyComponent extends Component {
-  render() {
-    const { classes, dataColumns } = this.props;
-    return (
-      <ListItem
-        classes={{ root: classes.item }}
-        divider={true}
-        secondaryAction={
-          <Box sx={{ root: classes.itemIconDisabled }}>
-            <MoreVertOutlined/>
-          </Box>
-        }
-      >
-        <ListItemIcon classes={{ root: classes.itemIconDisabled }}>
-          <Skeleton
-            animation="wave"
-            variant="circular"
-            width={30}
-            height={30}
-          />
-        </ListItemIcon>
-        <ListItemText
-          primary={
-            <div>
-              <div
-                className={classes.bodyItem}
-                style={{ width: dataColumns.relationship_type.width }}
-              >
-                <Skeleton
-                  animation="wave"
-                  variant="rectangular"
-                  width="90%"
-                  height="100%"
-                />
-              </div>
-              <div
-                className={classes.bodyItem}
-                style={{ width: dataColumns.entity_type.width }}
-              >
-                <Skeleton
-                  animation="wave"
-                  variant="rectangular"
-                  width="90%"
-                  height="100%"
-                />
-              </div>
-              <div
-                className={classes.bodyItem}
-                style={{ width: dataColumns.name.width }}
-              >
-                <Skeleton
-                  animation="wave"
-                  variant="rectangular"
-                  width={140}
-                  height="100%"
-                />
-              </div>
-              <div
-                className={classes.bodyItem}
-                style={{ width: dataColumns.created_at.width }}
-              >
-                <Skeleton
-                  animation="wave"
-                  variant="rectangular"
-                  width={140}
-                  height="100%"
-                />
-              </div>
-              <div
-                className={classes.bodyItem}
-                style={{ width: dataColumns.confidence.width }}
-              >
-                <Skeleton
-                  animation="wave"
-                  variant="rectangular"
-                  width={100}
-                  height="100%"
-                />
-              </div>
-            </div>
-          }
-        />
-      </ListItem>
-    );
-  }
-}
-
-SimpleStixObjectOrStixRelationshipStixCoreRelationshipLineDummyComponent.propTypes = {
-  dataColumns: PropTypes.object,
-  classes: PropTypes.object,
-};
-
-export const SimpleStixObjectOrStixRelationshipStixCoreRelationshipLineDummy = compose(
-  inject18n,
-  withStyles(styles),
-)(SimpleStixObjectOrStixRelationshipStixCoreRelationshipLineDummyComponent);
